Update folder structure (/src)
