+++ package.json
@@ -61,6 +61,7 @@
     "eslint": "^7.32.0",
     "eslint-plugin-import": "^2.24.2",
     "eslint-plugin-react": "^7.25.1",
+    "eslint-plugin-typescript-sort-keys": "^1.8.0",
     "stylelint": "^13.13.1",
     "stylelint-config-sass-guidelines": "^8.0.0"
   }
