+++ src/components/App/App.scss
@@ -155,4 +155,3 @@ body {
 * {
     box-shadow: none !important;
 }
-
