+++ src/components/App/index.tsx
@@ -2,7 +2,7 @@ import { ReactElement } from 'react'
 
 import { ChakraProvider } from '@chakra-ui/react'
 
-import { SetupProvider } from 'hooks/SetupContext'
+import { SetupProvider } from 'hooks/Setup'
 
 import Footer from 'components/Footer'
 import Main from 'components/Main'
