+++ src/index.tsx
@@ -2,10 +2,10 @@ import React from 'react'
 
 import ReactDOM from 'react-dom'
 
-import './index.css'
-import reportWebVitals from './reportWebVitals'
-
 import App from 'components/App'
+
+import './index.css'
+//import reportWebVitals from './reportWebVitals'
 
 ReactDOM.render(
   <React.StrictMode>
@@ -17,4 +17,4 @@ ReactDOM.render(
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals()
+//reportWebVitals()
